Add highlighting to commit diffs
